feat(components): ResourceList
