+++ packages/components/src/ResourcePicker/ResourcePicker.component.js
@@ -1,10 +1,12 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 import { getTheme } from '../theme';
 
 import ResourceList from '../ResourceList/ResourceList.component';
 
 import cssModule from './ResourcePicker.scss';
+import Toolbar from '../ResourceList/Toolbar';
 
 const theme = getTheme(cssModule);
 
@@ -13,7 +15,7 @@ export default function ResourcePicker(props) {
 		<div className={theme('tc-resource-picker')}>
 			<ResourceList
 				{...props}
-				className={theme('tc-resource-picker-list-container')}
+				className={theme('tc-resource-picker-list')}
 				toolbar={{ ...props.toolbar, nameFilerAsInput: true }}
 			/>
 		</div>
@@ -25,5 +27,10 @@ ResourcePicker.defaultProps = {
 };
 
 ResourcePicker.propTypes = {
-	...ResourceList.props,
+	className: PropTypes.string,
+	collection: PropTypes.arrayOf(PropTypes.object),
+	isLoading: PropTypes.bool,
+	onRowClick: PropTypes.func,
+	renderAs: PropTypes.func,
+	toolbar: PropTypes.shape(Toolbar.propTypes),
 };
